Separate authenticator (core) and its web version with http basic (UI)
