+++ ui/auth_test.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package auth
+package ui
 
 import (
 	"encoding/base64"
